refactor(review): split FactoryCacheCoordinator instrumentation helpers into partial
